infra/pyproject.toml
[tool.poetry]
name = "emr-config-builder-infra"
version = "0.0.1"
description = "IaC for EMR config Build"
authors = ["Denys Gonzaga <denys.gonzaga@proton.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.72.1"
pytest = "6.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
